[feat] 合并了 Upload 容器和 Backend 容器
